add compact cookie encoding.
Use exclusively Blake2s as a MAC and ChaCha20 as a stream cipher.
Blake2s is always faster than SHA2 and could be safely used as a MAC
without HMAC construction (therefore, it is much faster than HMAC-SHA256).
ChaCha20 is a bit slower than AES-CTR, but its usage causes less
allocations. And it is faster without AES hardware optimization
(appengine for example).